Bug 1176337 - Remove unnecessarily gcc-4.4-specific cast from nsTableRowFrame::ReflowChildren(), now that we require newer gcc. r=dholbert
UltraBlame original commit: fc4d79e8bad85bbf42a5c41c32603095da7cdbd3

+++ layout/tables/nsTableRowFrame.cpp
@@ -904,8 +904,7 @@ nsTableRowFrame::ReflowChildren(nsPresContext*           aPresContext,
         
         kidReflowState.emplace(aPresContext, aReflowState, kidFrame,
                                kidAvailSize,
-                               
-                               uint32_t(nsHTMLReflowState::CALLER_WILL_INIT));
+                               nsHTMLReflowState::CALLER_WILL_INIT);
         InitChildReflowState(*aPresContext, kidAvailSize, borderCollapse,
                              *kidReflowState);
